requirements.txt
youtube-transcript-api>=0.6.0
Flask
python-dotenv # Needed locally, Vercel uses its own env vars